requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
matplotlib>=3.7.0
aiosqlite>=0.19.0
